Add - 카드뷰 추가 - 카카오맵 마커 추가

+++ app/src/main/java/com/mungeun/gymforyou/domain/model/gym/Location.kt
@@ -5,6 +5,6 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Location(
-    val latitude: Int,
-    val longitude: Int
+    val latitude: Double,
+    val longitude: Double
 ) : Parcelable
